Refactor test execution
- Enable tests to run in paralallel
