+++ .prettierignore
@@ -1,3 +1,5 @@
 # Ignore artifacts:
 build
 coverage
+dist
+tsconfig.tsbuildinfo
