+++ jest.base.config.ts
@@ -1,0 +1,27 @@
+import type { Config } from "jest";
+
+const config: Config = {
+  clearMocks: true,
+  moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
+  moduleNameMapper: {
+    "^@fp-ts-playground/(.*)$": "<rootDir>/packages/$1/src",
+  },
+  preset: "ts-jest",
+  rootDir: "../../",
+  testEnvironment: "node",
+  testMatch: ["**/__tests__/**/*.ts", "**/?(*.)+(spec|test).ts"],
+  transform: {
+    "^.+\\.ts$": [
+      "ts-jest",
+      {
+        tsconfig: "tsconfig.json",
+      },
+    ],
+  },
+  projects: ["<rootDir>/packages/*/jest.config.ts"],
+  collectCoverage: true,
+  coverageDirectory: '<rootDir>/coverage',
+  coverageReporters: ['json', 'lcov', 'text', 'clover'],
+};
+
+export default config;
