--- jest.config.ts
@@ -1,26 +1,0 @@
-/**
- * For a detailed explanation regarding each configuration property, visit:
- * https://jestjs.io/docs/configuration
- */
-
-import type { Config } from "jest";
-
-const config: Config = {
-  clearMocks: true,
-  collectCoverage: true,
-  coverageDirectory: "coverage",
-  coverageProvider: "v8",
-  preset: "ts-jest",
-  testEnvironment: "node",
-  moduleNameMapper: {
-    "^@fp-ts-playground/(.*)$": "<rootDir>/packages/$1/src",
-  },
-  roots: ["<rootDir>/packages"],
-  transform: {
-    "^.+\\.tsx?$": "ts-jest",
-  },
-  testMatch: ["**/?(*.)+(spec|test).ts?(x)"],
-  moduleFileExtensions: ["ts", "tsx", "js", "jsx", "json", "node"],
-};
-
-export default config;
