+++ packages/application/jest.config.ts
@@ -1,0 +1,8 @@
+import { Config } from "jest";
+import baseConfig from "../../jest.base.config";
+
+const config: Config = {
+  ...baseConfig,
+};
+
+export default config;
