+++ packages/core/package.json
@@ -5,7 +5,8 @@
   "main": "dist/index.js",
   "types": "dist/index.d.ts",
   "scripts": {
-    "build": "tsc"
+    "build": "tsc",
+    "test": "jest"
   },
   "repository": "https://github.com/danielmunafo/fp-ts-playground.git",
   "author": "dann.tostes@gmail.com",
